requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
pathlib==1.0.1
python-dotenv==1.0.0